+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @renegade-fi/react
 
+## 0.5.1
+
+### Patch Changes
+
+- core: define and handle ws errors
+- Updated dependencies
+  - @renegade-fi/core@0.5.1
+
 ## 0.5.0
 
 ### Minor Changes